complete 2

+++ src/main/java/com/nttdata/bootcamp/assignement1/clients/infraestructure/CostumerRepository.java
@@ -4,11 +4,13 @@ import com.nttdata.bootcamp.assignement1.clients.model.Costumer;
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 /**
  * ReactiveMongoRepository nos facilita operaciones CRUD
  * El segundo parametro es Integer porque la id de Costumer es de tipo Integer
  */
 @Repository
-public interface CostumerRepository extends ReactiveMongoRepository<Costumer, Integer> {
+public interface CostumerRepository extends ReactiveMongoRepository<Costumer, BigInteger> {
 
 }
